Move email address regex to a common file.
